requirements.txt
Django>=5.0,<6
djangorestframework>=3.15
django-cors-headers>=4.3
drf-yasg>=1.21
requests>=2.31
